+++ tom_d4rt_ast/lib/src/runtime/callable.dart
@@ -1528,6 +1528,7 @@ class InterpretedFunction implements Callable {
                   throw RuntimeD4rtException(
                     "Error during bridged super constructor '$superConstructorName': ${e.message}",
                     originalStackTrace: e.originalStackTrace,
+                    originalException: e.originalException,
                   );
                 } catch (e, s) {
                   throw RuntimeD4rtException(
@@ -1710,6 +1711,7 @@ class InterpretedFunction implements Callable {
               throw RuntimeD4rtException(
                 "Error during implicit bridged super constructor: ${e.message}",
                 originalStackTrace: e.originalStackTrace,
+                originalException: e.originalException,
               );
             } catch (e, s) {
               throw RuntimeD4rtException(